+++ boot/src/main/java/io/gridgo/boot/data/jdbc/JdbcDataAccessHandler.java
@@ -1,7 +1,7 @@
-package io.gridgo.boot.data.support.impl;
+package io.gridgo.boot.data.jdbc;
 
 import io.gridgo.bean.BObject;
-import io.gridgo.boot.data.support.annotations.JdbcProduce;
+import io.gridgo.boot.data.support.impl.AbstractDataAccessHandler;
 import io.gridgo.framework.support.Message;
 import lombok.Getter;
